utils/requirements-dev.txt
aiohttp
kaggle
pandas
pandas-stubs
python-dotenv
requests
mypy
pytest
flake8
isort
black